Seeker setting, cursor setting, zooming in and out, pausing, replaying, playing, stopping | all these things now work with multiple blocks
