+++ client/reducers/workspace.js
@@ -16,7 +16,7 @@ export default handleActions({
   },
 
   SET_SEEKER: (state, action) => {
-    return {...state, timing: {...state.timing, seeker: action.payload}};
+    return {...state, timing: {...state.timing, seeker: Math.max(0,action.payload)}};
   },
 
   SET_ZOOM: (state, action) => {
@@ -30,7 +30,7 @@ export default handleActions({
   },
 
   SET_CURSOR: (state, action) => {
-    return {...state, timing: {...state.timing, cursor: action.payload}};
+    return {...state, timing: {...state.timing, cursor: Math.max(0,action.payload)}};
   },
 
   SET_SPEED: (state, action) => {
