+++ client/store/store.js
@@ -31,3 +31,5 @@ store.dispatch(workspaceActions.setCursor(0));
 reduxRouter.listenForReplays(store);
 
 export default store;
+
+// ab561a8c
