+++ models/defaultWorkspace.js
@@ -3,11 +3,18 @@ module.exports = {
   rows: [{
     rowId: 0,
     rawAudio: 'default.ogg',
-    audioBlocks: [{
-      file_end: undefined,
-      row_offset: 0,
-      file_offset: 0,
-      flags: []
-    }]
+    audioBlocks: [
+      { 
+        file_end: 264,
+        row_offset: 0,
+        file_offset: 0,
+        flags: [] 
+      }, {
+        row_offset: 132,
+        file_offset: 264,
+        file_end: undefined,
+        flags: []
+      }
+    ]
   }]
 }
